Enabled as much MacOS compilable plugins as possible

+++ plugins/src/tascar_ap_feedbackdelay.cc
@@ -8,7 +8,7 @@ public:
   void add_variables( TASCAR::osc_server_t* srv );
   ~feedbackdelay_t();
 private:
-  size_t maxdelay;
+  uint64_t maxdelay;
   double f;
   double feedback;
   TASCAR::varidelay_t* dl;
